Feat/advanced help modal markdown (#644)
* Use the tailwind typography plugin to render markdown

Without this bullet points and links will not render correctly because
tailwind by default removes styling on all html tags

* Increase max width of dialog to 3xl

The size of the dialog was too small. So I updated it to be 3xl to be
larger if the window is large enough for it. It will still resize
responsively if the window is shrunk

* Fix lint issue